Refactor productivity page into separate components
